Renames.
tested_algorithms -> algorithms
