+++ algorithm_tester/helpers.py
@@ -1,5 +1,5 @@
 import os
-from algorithm_tester.tested_algorithms import Algorithms
+from algorithm_tester.algorithms import Algorithms
 from algorithm_tester.tester_dataclasses import AnalysisFile
 
 class FilePair:
